update: modularize login with LNURL and logout
